requirements.txt
SpeechRecognition==3.10.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
python-dotenv==1.0.0
pyaudio==0.2.13
numpy==1.24.3
sounddevice==0.4.6
scipy==1.10.1
python-dateutil==2.8.2
pvporcupine
tzlocal
pyttsx3 